+++ azure-pipelines/Master/BuildAndPushPackage.yml
@@ -57,6 +57,7 @@ parameters:
     54: FGS.Extensions.Hosting.Middleware
     55: FGS.Extensions.Hosting.DependencyInjection.Autofac
     56: FGS.Extensions.Hosting.Logging.Serilog
+    57: FGS.Tests.Support.Extensions.Hosting.Middleware.Mocking
 
 stages:
